Separate js/main.js to modules
